requirements.txt
# IDP System v2.0 Requirements
# Upgraded with Hybrid Search, Reranking, and Local LLMs

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# OCR & Vision (High DPI)
paddlepaddle-gpu==2.6.0  # Use paddlepaddle for CPU-only
paddleocr==2.7.0
pdf2image==1.16.3
PyMuPDF==1.23.8
Pillow==10.2.0
opencv-python==4.9.0.80

# Layout Analysis
layoutparser==0.3.4
detectron2  # Install from: pip install 'git+https://github.com/facebookresearch/detectron2.git'

# Embeddings
sentence-transformers==2.3.1
torch>=2.1.0

# Vector Store (Qdrant)
qdrant-client==1.7.0

# Reranking (NEW)
# Uses sentence-transformers CrossEncoder

# LLM Clients
anthropic==0.18.1
openai==1.12.0
httpx==0.26.0  # For Ollama/vLLM calls

# Schema Validation
jsonschema==4.21.1

# Caching (Optional)
redis==5.0.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
python-dotenv==1.0.1
tqdm==4.66.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
isort==5.13.2
